Specify connection names with symbol instead of string
Down in the guts of AR, it uses the class of the connection name
parameter to determine how the parameters are configured.  If a string
is passed in, AR assumes that the string is a URL and tries to pull it
apart into its parts.  If the name is a symbol, though, it assumes
that the params are in a hash, which is what we want.